feat: Implement a basic lighting system with `LightManager` and `LightComponent`, integrating it into the rendering pipeline

+++ Vulkan_Engine/Editor/EditorCamera.h
@@ -10,7 +10,7 @@
 #include "ui/SceneUi.h"
 #include "gizmos/ImGuizmo.h"
 
-extern float deltaTime;
+extern double deltaTime;
 
 class EditorCamera  
 {
